make updateMethod of main and layer-list part of first-update
