+++ elements/layercontrol/src/methods/layercontrol/first-update.js
@@ -1,12 +1,24 @@
 /**
  * Retrieves the EOxMap element based on the provided selector in EOxLayerControl's 'for' property.
+ * Updates the EOxLayerControl's map property based on the target element's map.
  *
  * @param {import("../../main").EOxLayerControl} EOxLayerControl - The EOxLayerControl element.
  * @returns {import("@eox/map/main").EOxMap | null} The EOxMap element or null if not found.
  */
 const firstUpdatedMethod = (EOxLayerControl) => {
   const mapSelector = EOxLayerControl.for;
-  return document.querySelector(mapSelector);
+
+  // Find the element specified in EOxLayerControl.for
+  /** * @type {Element & { map: import("ol").Map }} */
+  const foundElement = document.querySelector(mapSelector);
+
+  // Check if the element exists and its map property differs from EOxLayerControl.map
+  if (foundElement && foundElement.map !== EOxLayerControl.map) {
+    // Update the EOxLayerControl's map property
+    EOxLayerControl.map = foundElement.map;
+  }
+
+  return foundElement;
 };
 
 export default firstUpdatedMethod;
